+++ src/quant.c
@@ -1351,7 +1351,7 @@ sixel_quant_apply_palette(
                 }
             }
         }
-        memcpy(palette, new_palette, *ncolors * depth);
+        memcpy(palette, new_palette, (size_t)(*ncolors * depth));
     } else {
         for (y = 0; y < height; ++y) {
             for (x = 0; x < width; ++x) {
